termux/requirements.txt
aiogram==3.*
aiosqlite==0.20.*
python-dotenv==1.*
telethon==1.*
